refactored unit test
